remove utils::client
