+++ integration-test/src/help.rs
@@ -3,13 +3,13 @@ use super::const_definition::{
     PW_LOCK_DEVNET_TYPE_HASH, RPC_TRY_COUNT, RPC_TRY_INTERVAL_SECS, SCRIPT_CONFIG,
 };
 use crate::const_definition::CURRENT_OTX_POOL_SERVICE_PROCESS;
+use crate::utils::client::ckb_client::CkbRpcClient;
 use crate::utils::client::mercury_client::MercuryRpcClient;
 use crate::utils::instruction::command::run_command_spawn;
 use crate::utils::instruction::{ckb::generate_blocks, ckb::unlock_frozen_capacity_in_genesis};
 
 use otx_pool_config::{parse, AppConfig, ConfigFile};
 use otx_util_client::OtxPoolRpcClient;
-use utils::client::ckb_client::CkbRpcClient;
 
 use common::lazy::{
     ACP_CODE_HASH, CHEQUE_CODE_HASH, DAO_CODE_HASH, PW_LOCK_CODE_HASH, SECP256K1_CODE_HASH,
