+++ util/client/src/otx_pool_client.rs
@@ -1,8 +1,8 @@
 use super::{request, RpcClient};
 
-use otx_pool_plugin_atomic_swap::SwapProposalWithOtxId;
 use otx_format::{jsonrpc_types::OpenTransaction, types::OpenTxStatus};
 use otx_plugin_protocol::PluginInfo;
+use otx_pool_plugin_atomic_swap::SwapProposalWithOtxId;
 
 use anyhow::Result;
 use ckb_types::H256;
