+++ utils/Cargo.toml
@@ -14,10 +14,8 @@ ckb-sdk = "=2.3.0"
 ckb-types = "0.105"
 faster-hex = "0.6"
 log = "0.4"
-jsonrpc-core = "18.0"
 once_cell = "1.10"
 rand = "0.8"
-reqwest = { version = "0.11", features = ["blocking", "json"] }
 secp256k1 = { version = "0.24", features = ["recovery"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
