+++ utils/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod aggregator;
 pub mod build_tx;
-pub mod client;
